+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: kzarins <kzarins@student.42heilbronn.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/09 11:28:16 by kzarins           #+#    #+#             */
-/*   Updated: 2024/10/09 18:33:02 by kzarins          ###   ########.fr       */
+/*   Updated: 2024/10/12 20:47:14 by kzarins          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
@@ -14,7 +14,7 @@
 // The original strlen function does not check if a 
 // NULL pointer is given and it crashes!
 // ft_strlen(NULL) == 0;
-int	ft_strlen(const char *str)
+size_t ft_strlen(const char *str)
 {
 	int	len;
