+++ ft_strnstr.c
@@ -6,14 +6,14 @@
 /*   By: kzarins <kzarins@student.42heilbronn.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/09 12:32:31 by kzarins           #+#    #+#             */
-/*   Updated: 2024/10/11 16:00:22 by kzarins          ###   ########.fr       */
+/*   Updated: 2024/10/12 20:38:33 by kzarins          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "libft.h"
 
-int	ft_strcmp(const char *str1, const char *str2, int len);
+int	ft_strcmp(const char *str1, const char *str2, size_t len);
 
-char	*ft_strnstr(const char *heystack, const char *needle, unsigned int len)
+char	*ft_strnstr(const char *heystack, const char *needle, size_t len)
 {
 	if (*needle == '\0')
 		return ((char *)heystack);
@@ -27,7 +27,7 @@ char	*ft_strnstr(const char *heystack, const char *needle, unsigned int len)
 	return (0);
 }
 
-int	ft_strcmp(const char *str1, const char *str2, int len)
+int	ft_strcmp(const char *str1, const char *str2, size_t len)
 {
 	while (*str1 && *str2 && len > 0)
 	{
